week-07/requirements.txt
grpcio==1.68.0
grpcio-tools==1.68.0
protobuf==5.28.0
setuptools